+++ src/compute_image.hpp
@@ -8,4 +8,4 @@ struct Image {
 
 void free_image(Image* image);
 
-Image* compute_image_c(const unsigned int* filled_pixels, unsigned int n_filled_pixels, unsigned int width, unsigned int height);
+Image* compute_image_c(const long* filled_pixels, unsigned int n_filled_pixels, unsigned int width, unsigned int height);
